// src/types/shared.ts
export type LinkIconType = 'auto' | 'library' | 'file' | 'url'
// ...
export interface LinkElem {
    _id: string
    parent?: string
    folder?: false
    order: number
    title: string
    url: string
    icon?: LinkIcon
}

export interface LinkIcon {
    type: LinkIconType
    value?: string
}